% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_pcon_names.R
\name{clean_pcon_names}
\alias{clean_pcon_names}
\title{Clean British Parliamentary Constituency Names}
\usage{
clean_pcon_names(x)
}
\arguments{
\item{x}{A vector of constituency names.}
}
\value{
A vector of cleaned constituency names.
}
\description{
British parliamentary constituencies are not often coded in a consistent manner. Fixing this can also be very time-consuming. This function uses the constituencies dataset in the package to standardise a vector of constituencies names.
}
\examples{
clean_pcon_names("Ynys Môn")
}
