% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{constituencies}
\alias{constituencies}
\title{Historic Parliamentary Constituencies}
\format{
A data frame:
\describe{
\item{name}{Name of parliamentary constituency}
\item{gss_code}{Government Statistical Service code}
\item{start}{Date constituency was introduced}
\item{end}{Date constituency was retired}
}
}
\usage{
constituencies
}
\description{
A dataset containing the names and GSS codes of all historic constituencies since 1918.
}
\keyword{datasets}
