% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{election_results}
\alias{election_results}
\title{Historic Election Results by Constituency, 1929 onwards}
\format{
A data frame:
\describe{
\item{date}{Election date}
\item{constituency}{Name of constituency}
\item{seats}{Number of seats in constituency}
\item{region}{Region or country}
\item{electorate}{Size of constituency electorate}
\item{con_votes}{Number of Conservative votes}
\item{lab_votes}{Number of Labour votes}
\item{lib_votes}{Number of Liberal votes}
\item{nat_votes}{Number of Nationalist votes in Scotland and Wales}
\item{oth_votes}{Number of votes for any other party}
}
}
\usage{
election_results
}
\description{
A dataset containing historic constituency-level election results in Britain. Taken from the House of Commons library https://commonslibrary.parliament.uk/research-briefings/cbp-8647/.
}
\keyword{datasets}
