% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit Poll Pooling Model}
\usage{
fit_model(data, init, final, party, alpha_init, alpha_final)
}
\arguments{
\item{data}{Raw pollbase data.}

\item{init}{Initial election date.}

\item{final}{Final election date.}

\item{party}{A string reflecting a party. Either "con", "lab", or "lib".}

\item{alpha_init}{The party's vote share at the initial election.}

\item{alpha_final}{The party's vote share at the final election.}
}
\description{
This function provides a simple way to fit a variant of Simon Jackman's "polling the polls" model.
}
