% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last_election.R
\name{get_last_election}
\alias{get_last_election}
\title{Get Previous Election Based on Date}
\usage{
get_last_election(x)
}
\arguments{
\item{x}{A vector of dates.}
}
\description{
Given a vector of dates, this function provides a vector of the same length that corresponds to the date of the last election for each element in the vector. Applies only to the 1935 election onwards.
}
