% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_result.R
\name{get_result}
\alias{get_result}
\title{Get Election Result Based on Party and Election}
\usage{
get_result(election = "1997-05-01", party = "lab")
}
\arguments{
\item{election}{A date corresponding to a UK election.}

\item{party}{A string reflecting a party. Either "con", "lab", or "lib".}
}
\description{
Given a vector of dates, this function provides a vector of the same length that corresponds to the date of the last election for each element in the vector. Applies only to the 1935 election onwards.
}
