% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_results}
\alias{load_results}
\title{Load historic election results from the House of Commons library}
\usage{
load_results()
}
\description{
This function downloads and loads historic elections data.
}
