% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/party_colours.R
\name{party_colours}
\alias{party_colours}
\title{British Political Party Colours}
\usage{
party_colours(x)
}
\arguments{
\item{x}{A vector of party names.}
}
\value{
A vector of corresponding hex codes
}
\description{
British parliamentary constituencies are not often coded in a consistent manner. Fixing this can also be very time-consuming. This function uses the constituencies dataset in the package to standardise a vector of constituencies names.
}
\examples{
party_colours(c("Ynys Môn", "Derby North", "North, Derby"))
}
