% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollbasepro.R
\docType{data}
\name{pollbasepro}
\alias{pollbasepro}
\title{Daily Estimates of British Voting Intention, 1955 to Present}
\format{
A data frame:
\describe{
\item{date}{Date}
\item{election}{Date of last general election}
\item{govt}{Largest party in government after the last general election}
\item{con_est}{Posterior mean: Conservative voting intention}
\item{con_err}{Posterior error: Conservative voting intention}
\item{lab_est}{Posterior mean: Labour voting intention}
\item{lab_err}{Posterior error: Labour voting intention}
\item{lib_est}{Posterior mean: Liberal voting intention}
\item{lib_err}{Posterior error: Liberal voting intention}
\item{con_ldr}{Leader of the Conservative Party}
\item{lab_ldr}{Leader of the Labour Party}
\item{lib_ldr}{Leader of the Liberals (various forms)}
\item{week}{Weekly subset indicator}
\item{month}{Monthly subset indicator}
\item{quarter}{Quarterly subset indicator}
\item{year}{Yearly subset indicator}
}
}
\usage{
pollbasepro
}
\description{
A dataset containing daily estimates of aggregate voting intention in Britain for each of the country's three largest parties between 1955 and the present.
}
\keyword{datasets}
