% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swing.R
\name{swing}
\alias{swing}
\title{Compute Two-Party Swing}
\usage{
swing(party_a = NULL, party_b = NULL, type = "butler")
}
\arguments{
\item{party_a}{A vector of length 2. The first value corresponds to the most recent election.}

\item{party_b}{A vector of length 2. The first value corresponds to the most recent election.}

\item{type}{Which swing calculation to use. Either "butler" or "steed". Defaults to "butler".}
}
\description{
This convenience function takes two numeric vectors of vote share data and returns the estimated two-party swing.
}
\examples{
\dontrun{
swing(party_a = c(0.5, 0.4), party_b = c(0.5, 0.6), type = "butler")
}
}
