% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_new_polls.R
\name{update_pollbasepro}
\alias{update_pollbasepro}
\title{Update Voting Intention Polls, 2019-Present}
\usage{
update_pollbasepro()
}
\description{
This function updates PollBasePro to include the latest polling data and outputs Twitter content.
}
